Adding DTOs that somehow didn't make it in the last commit.

+++ src/API/DTO/BanDTO.java
@@ -17,8 +17,8 @@ import lombok.ToString;
 public class BanDTO {
 
     @JsonProperty("pickTurn")
-    private Long pickTurn;
+    private Integer pickTurn;
     @JsonProperty("championId")
-    private Long championId;
+    private Integer championId;
 
 }
